refactor: add public API barrels to db layer and enforce import boundary
Route all external db imports through index.ts (types) and
index.server.ts (runtime) instead of reaching into internal files.
ESLint now blocks direct imports of db internals from outside app/db/.
