+++ app/db/index.server.ts
@@ -1,0 +1,3 @@
+export { getDb } from "./client.server";
+export { users, drinks } from "./schema";
+export { resetAndSeedDatabase } from "./reset.server";
