+++ app/db/index.ts
@@ -1,0 +1,1 @@
+export type { User, NewUser, Drink, NewDrink } from "./schema";
