+++ app/modules/auth/implementation/types.ts
@@ -1,4 +1,4 @@
-import type { User } from "#/app/db/schema";
+import type { User } from "#/app/db";
 
 export type AuthenticatedUser = {
   id: User["id"];
