+++ app/modules/search/implementation/search.server.ts
@@ -1,7 +1,7 @@
 import { remember, forget } from "@epic-web/remember";
 import MiniSearch from "minisearch";
 import { getPublishedDrinks } from "#/app/modules/drinks/index.server";
-import type { Drink } from "#/app/db/schema";
+import type { Drink } from "#/app/db";
 
 const CACHE_KEY = "minisearch-index";
