+++ test/setup.ts
@@ -1,6 +1,6 @@
 import { afterAll, afterEach, beforeAll } from "vitest";
 import { migrate } from "drizzle-orm/better-sqlite3/migrator";
-import { getDb } from "#/app/db/client.server";
+import { getDb } from "#/app/db/index.server";
 import { server } from "./server";
 
 migrate(getDb(), { migrationsFolder: "./drizzle" });
